chore: NestedTable its own component + sub table sorting...
